requirements.txt
Django==6.0.3
djangorestframework==3.17.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
black==26.3.1
isort==8.0.1
flake8==7.3.0
gunicorn==25.3.0
